clearned code and restored migrations
